raspbberrypi_send_to_iothub/requirements.txt
pyserial
azure-eventhub
python-dotenv